Hugely condensed the URL-matching regex in #'aff-replace-urls-in-region; it now passes its unit test.
This simplification makes it readable at the cost of matching fewer URLs (e.g., URLs that include characters that can't be encoded in ASCII will not be matched, and most URLs that end with a ")" will not be correctly matched). However, as far as I can tell, all of the URLs that this package cares about matching, limit themselves to the ASCII realm; so this package will still find all URLs relevant to its problem domain.